Refactored simulator actions
