+++ src/components/RegistryView.jsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useState} from "react";
 import {RUST} from "../rust_functions.js";
-import {CALLBACK_TYPES, CALLBACKS_TRIGGERS} from "../App.jsx";
+import {CALLBACK_TYPES, CALLBACKS_TRIGGERS} from "./simulator.js";
 
 function RegistryView({loaded, registerCallback}) {
     const [registers, setRegisters] = useState([]);
